+++ src/features/venueMap/components/VenueMapView/VenueMapView.web.tsx
@@ -1,0 +1,1 @@
+export const VenueMapView = () => null
