+++ src/features/venueMap/pages/VenueMap/VenueMap.web.tsx
@@ -1,0 +1,3 @@
+import { VenueMapBase } from 'features/venueMap/pages/VenueMap/VenueMapBase'
+
+export const VenueMap = VenueMapBase
